requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
pydantic>=2.7
python-multipart>=0.0.9
passlib[bcrypt]>=1.7
python-jose>=3.3
python-dotenv>=1.0
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic-settings==2.4.0

SQLAlchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2

python-jose==3.3.0
passlib==1.7.4  # PBKDF2-SHA256 (sin binarios)

redis==5.0.7
fastapi-limiter==0.1.6
